+++ client/potentials/Morse/Morse.cpp
@@ -42,7 +42,7 @@ void Morse::setParameters(double De, double a, double re, double cutoff)
 // pointer to number of atoms, pointer to array of positions	
 // pointer to array of forces, pointer to internal energy
 // adress to supercell size
-void Morse::force(long N, const double *R, const long *, double *F, double *U, const double *box)
+void Morse::force(long N, const double *R, const int *, double *F, double *U, const double *box)
 {
       assert((box[0] > 0) and (box[1] > 0) and (box[2] > 0));
       double diffR=0, diffRX, diffRY, diffRZ;
