+++ client/potentials/bopfox/bopfox.h
@@ -11,11 +11,11 @@ class bopfox : public PotentialsInterface
 		~bopfox();
         void initialize() {};
         void cleanMemory(void);    
-        void force(long N, const double *R, const long *atomicNrs, double *F, double *U, const double *box);
+        void force(long N, const double *R, const int *atomicNrs, double *F, double *U, const double *box);
 
 
     private:
-        void writeFOX(long N, const double *R, long const *atomicNrs, const double *box);
+        void writeFOX(long N, const double *R, int const *atomicNrs, const double *box);
         void readFU(long N, double *F, double *U);
 
 };
